requirements.txt
fastapi==0.79.0
uvicorn==0.18.2

PyJWT==2.4.0
python-decouple==3.6